+++ CHANGELOG.md
@@ -10,6 +10,8 @@ since `1.7.0`.
 ### Added
 - Added `TCOD_context_convert_event_coordinates` and `TCOD_Context::convert_event_coordinates` as an easier way to
   convert mouse pixel coordinates to tile coordinates.
+- Several C++ drawing and print functions have been made public.
+- Several C++ tileset functions have been made public.
 
 ### Deprecated
 - Deprecate `TCOD_chars_t` enum values because they are non-Unicode.